Replace / with //.
Python3.12 does not like a single / in the docstring and throws syntax
warnings about it. We follow the document and replace all / with //.

We dont know if any / where is should be a / are mistakenly changed to //, like some latex commands in plot.py. We need more tests to
figure them out.